// src/api-client/model/itemRequest.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * 빌릴게(Billilge) API
 * 국민대학교 소프트웨어융합대학 복지물품 대여 서비스
 * OpenAPI spec version: v1
 */
import type { ItemRequestType } from './itemRequestType';

export interface ItemRequest {
  /** 물품 이름 */
  name: string;
  /** 물품 타입 */
  type: ItemRequestType;
  /** 물품 개수 */
  count: number;
}