// src/api-client/model/itemRequestType.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * 빌릴게(Billilge) API
 * 국민대학교 소프트웨어융합대학 복지물품 대여 서비스
 * OpenAPI spec version: v1
 */

/**
 * 물품 타입
 */
export type ItemRequestType = typeof ItemRequestType[keyof typeof ItemRequestType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ItemRequestType = {
  CONSUMPTION: 'CONSUMPTION',
  RENTAL: 'RENTAL',
} as const;